+++ setup.cfg
@@ -21,4 +21,3 @@ markers =
 [coverage:run]
 omit =
   cola/*_fns.py
-  cola/utils/test_utils.py
