+++ tests/algorithms/test_iram.py
@@ -3,14 +3,14 @@ import numpy as np
 from cola.backends import all_backends
 from cola.fns import lazify
 from cola.linalg.eig.iram import iram
-from cola.utils.test_utils import (
+from cola.utils.utils_for_tests import (
     generate_pd_from_diag,
     generate_spectrum,
-    get_numpy_dtype,
     get_xnp,
     parametrize,
     relative_error,
 )
+from cola.utils.utils_linalg import get_numpy_dtype
 
 
 @parametrize(all_backends)
