+++ tests/algorithms/test_lobpcg.py
@@ -3,7 +3,7 @@ import numpy as np
 from cola.backends import all_backends
 from cola.fns import lazify
 from cola.linalg.eig.lobpcg import lobpcg
-from cola.utils.test_utils import generate_pd_from_diag, generate_spectrum, get_xnp, parametrize, relative_error
+from cola.utils.utils_for_tests import generate_pd_from_diag, generate_spectrum, get_xnp, parametrize, relative_error
 
 
 @parametrize(all_backends)
