+++ tests/algorithms/test_qr.py
@@ -3,7 +3,7 @@ import numpy as np
 from cola.backends import tracing_backends
 from cola.linalg.tbd.qr import shifted_qr
 from cola.ops import Dense
-from cola.utils.test_utils import generate_pd_from_diag, generate_spectrum, get_xnp, parametrize, relative_error
+from cola.utils.utils_for_tests import generate_pd_from_diag, generate_spectrum, get_xnp, parametrize, relative_error
 
 
 @parametrize(tracing_backends)
