+++ tests/linalg/test_solve.py
@@ -1,9 +1,9 @@
 from cola.annotations import PSD
 from cola.linalg.inverse.inv import solve
 from cola.fns import lazify
-from cola.utils.test_utils import get_xnp, parametrize, relative_error
+from cola.utils.utils_for_tests import get_xnp, parametrize, relative_error
 from cola.backends import all_backends
-from cola.utils.test_utils import generate_spectrum, generate_pd_from_diag
+from cola.utils.utils_for_tests import generate_spectrum, generate_pd_from_diag
 
 
 @parametrize(all_backends)
