+++ tests/linalg/test_sqrt.py
@@ -4,9 +4,9 @@ from cola.linalg import sqrt
 from cola.linalg.algorithm_base import Auto
 from cola.ops import Diagonal
 from cola.annotations import SelfAdjoint
-from cola.utils.test_utils import get_xnp, parametrize, relative_error
+from cola.utils.utils_for_tests import get_xnp, parametrize, relative_error
 from cola.backends import all_backends
-from cola.utils.test_utils import generate_spectrum, generate_pd_from_diag
+from cola.utils.utils_for_tests import generate_spectrum, generate_pd_from_diag
 
 _tol = 1e-6
